+++ project_Backend/packages_table.sql
@@ -7,10 +7,10 @@ DROP TABLE IF EXISTS package_sizes;
 CREATE TABLE package_sizes(
 
     package_type int NOT NULL UNIQUE,
-    length float,
-    width float,
-    height float,
-    volume float,
+    length decimal,
+    width decimal,
+    height decimal,
+    volume decimal,
 
     PRIMARY KEY(package_type)
